refactor: added registry and subpackages for subjects
