+++ bfm/subject/registry.py
@@ -1,0 +1,6 @@
+"""Contains registries for `subjects` and `datasets`."""
+from bfm.core.registry import Registry
+
+
+subjects = Registry("subjects")
+datasets = Registry("datasets")
