+++ bfm/subject/subjects/ajile12.py
@@ -5,7 +5,8 @@ import numpy as np
 import torch
 from pynwb import NWBHDF5IO
 
-from subject.subject import Subject
+from bfm.subject.base import Subject
+from bfm.subject.registry import subjects
 
 from dotenv import load_dotenv
 load_dotenv()  # Load environment variables from .env file
@@ -16,6 +17,7 @@ warnings.filterwarnings("ignore", message="Ignoring cached namespace .* because 
 AJILE_ROOT_DIR = os.environ["AJILE_ROOT_DIR"]
 os.environ["HDF5_USE_FILE_LOCKING"] = "FALSE" # Disable file locking for HDF5 files. This is helpful for parallel processing.
 
+@subjects.register("ajile12")
 class AjileSubject(Subject):
     """ 
         This class is used to load the neural data for a given subject and trial.
